requirements.txt
# 基础依赖
openai>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0
PyPDF2>=3.0.0
python-docx>=0.8.11
asyncio>=3.4.3
aiofiles>=23.0.0
rich>=13.0.0
click>=8.0.0
loguru>=0.7.0
tiktoken>=0.5.0
nltk>=3.8.0
spacy>=3.7.0
langdetect>=1.0.9
sentence-transformers>=2.2.0
chromadb>=0.4.0
tenacity>=8.0.0
jinja2>=3.1.0
pymupdf
pathlib
datetime
dataclasses
enum
json
re
typing

# 新增：高级AI和深度学习依赖
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
networkx>=3.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 新增：创造性思维和推理依赖
nltk>=3.8
spacy>=3.6.0
wordnet
conceptnet-lite
jellyfish>=0.11.0

# 新增：专家系统依赖
owlready2>=0.44
rdflib>=6.3.0
SPARQLWrapper>=2.0.0

# 新增：性能和监控
psutil>=5.9.0
memory-profiler>=0.61.0
line-profiler>=4.0.0

# 新增：并行计算
joblib>=1.3.0
dask>=2023.5.0
ray>=2.5.0

# 新增：可视化和分析
plotly>=5.15.0
dash>=2.11.0
graphviz>=0.20.0

# 新增：文本处理和NLP
textblob>=0.17.0
polyglot>=16.7.4
langdetect>=1.0.9
emoji>=2.5.0 